Separate REST endpoints to it's own class
